Method to get the next shortest job
Will be called by each CPU thread once they're free. I added an empty queue handle, but I think when I make the loop telling CPU's to access the queue, the condition will be that the queue is not empty. This of course breaks if the Nodes have really long arrival times, and the queue is meant to sit empty, but ehhh that seems like a really edgy edge-case to handle