Setting up the actual simulation loop
I was trying to do this in an event-driven simulation way because I learned in Software Engineering that it is a bit more efficient. However, trying to think of how I want to juggle all of the math for calculating current_time is making my head a little sore. So for now I am doing the easy solution of just having an event loop and incrementing time by one unit constantly until the next arrival happens or otherwise.