Made logic clearer
Because I wasn't using the other dequeue, I just remade enqueue so that it is ordered by the shortest job, which means I can just dequeue from the queue instead of finding the next shortest job. More efficient and clearer. However, that is really me coping because I can't fix the seg fault :)